+++ src/pages/LayoutPage.tsx
@@ -1,7 +1,7 @@
 import Header from '../components/layout/Header'
 import Footer from '../components/layout/Footer'
 import { Outlet, useLocation } from 'react-router'
-import { NotiBadge } from '../components/layout/NotiBadge'
+import { NotiBadge } from '../components/layout/Notification/NotiBadge'
 
 function LayoutPage() {
   const location = useLocation()
